requirements.txt
absl-py==2.3.1
accelerate==1.11.0
addict==2.4.0
aenum==3.1.16
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.2
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
basicsr==1.4.2
beautifulsoup4==4.14.2
bleach==6.3.0
blendmodes==2025
blessings==1.7
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
clean-fid==0.1.35
click==8.3.1
clip==0.2.0
cloudpickle==3.1.2
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.2.3
compel==2.2.1
contourpy==1.3.2
coverage==7.12.0
cycler==0.12.1
Cython==3.2.3
dctorch==0.1.2
debugpy==1.8.17
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.35.2
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
facexlib==0.3.0
fastapi==0.124.2
fasteners==0.20
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
filterpy==1.4.5
flash_attn==2.8.3
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gfpgan @ git+https://github.com/Disty0/GFPGAN@ae0f7e44fafe0ef4716f3c10067f8f379b74c21c
gitdb==4.0.12
GitPython==3.1.32
glcontext==3.0.0
gradio==6.2.0
gradio_client==2.0.2
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
igraph==1.0.0
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
invisible-watermark==0.2.0
ipykernel==7.1.0
ipython==8.37.0
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonmerge==1.8.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
kornia==0.6.7
kornia_rs==0.1.10
lark==1.3.1
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.44.0
lmdb==1.7.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
moderngl==5.12.0
mpmath==1.3.0
multidict==6.7.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nodeenv==1.9.1
notebook==7.5.0
notebook_shim==0.2.4
numba==0.61.2
numexpr==2.11.0
numpy==1.24.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.2.3
onnx==1.19.1
onnxruntime==1.23.2
open-clip-torch==2.20.0
opencv-contrib-python==4.12.0.88
opencv-contrib-python-headless==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
optimum==2.0.0
orderly-set==5.5.0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
patch-ng==1.19.0
peft==0.17.1
pexpect==4.9.0
pi_heif==0.22.0
piexif==1.1.3
pillow==10.4.0
pillow-avif-plugin==1.4.3
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pylint==4.0.3
PyMatting==1.1.14
pymeshfix==0.17.1
pyparsing==3.2.3
pyproject-api==1.10.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytorch-lightning==1.9.0
pytz==2025.2
pyvista==0.46.4
PyWavelets==1.8.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
rembg==2.0.68
requests==2.32.4
resize-right==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rootpath==0.1.1
rpds-py==0.29.0
ruff==0.14.6
safehttpx==0.1.7
safetensors==0.6.2
scikit-image==0.21.0
scipy==1.14.1
scooby==0.11.0
seam-carving==1.1.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spandrel==0.3.4
spandrel_extra_arches==0.1.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
sympy==1.13.1
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
tensordict==0.8.3
termcolor==3.2.0
terminado==0.18.1
texttable==1.7.0
tifffile==2025.5.10
timm==1.0.16
tinycss2==1.4.0
tokenizers==0.22.1
tomesd==0.1.3
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.5.1+cu121
torchdiffeq==0.2.3
torchmetrics==1.8.2
torchsde==0.2.6
torchvision==0.20.1+cu121
tornado==6.5.2
tox==4.32.0
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.57.1
trimesh==4.9.0
triton==3.1.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
udf_ext==0.0.0
uri-template==1.3.0
urllib3==1.26.19
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@e0eccbc508045e374c9ff48c8fbc21611723bcd0
uvicorn==0.38.0
virtualenv==20.35.4
voluptuous==0.15.2
vtk==9.5.2
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==11.0.3
Werkzeug==3.1.3
xformers==0.0.29.post1
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
